+++ src/AvaloniaSyncer/Sections/NewSync/LeftOnlyFileActionViewModel.cs
@@ -1,0 +1,61 @@
+﻿using System;
+using System.Reactive;
+using System.Reactive.Linq;
+using System.Reactive.Subjects;
+using System.Threading;
+using System.Threading.Tasks;
+using CSharpFunctionalExtensions;
+using Zafiro.Actions;
+using Zafiro.CSharpFunctionalExtensions;
+using Zafiro.FileSystem;
+using Zafiro.FileSystem.Actions;
+using Zafiro.UI;
+
+namespace AvaloniaSyncer.Sections.NewSync;
+
+internal class LeftOnlyFileActionViewModel : IFileActionViewModel
+{
+    private readonly BehaviorSubject<LongProgress> progress = new(new LongProgress());
+
+    public LeftOnlyFileActionViewModel(ZafiroPath left, IZafiroDirectory source, IZafiroDirectory destination)
+    {
+        Left = left;
+        Source = source;
+        Destination = destination;
+        Sync = StoppableCommand.Create(() => Observable.FromAsync(ct => OnSync(source, destination, ct)), Maybe<IObservable<bool>>.None);
+        IsSyncing = Sync.IsExecuting;
+    }
+
+    public ZafiroPath Left { get; }
+    public IZafiroDirectory Source { get; }
+    public IZafiroDirectory Destination { get; }
+
+    public IObservable<bool> IsSyncing { get; }
+
+    public StoppableCommand<Unit, Result> Sync { get; }
+
+    public string Description => $"Copy {Left}";
+    public IObservable<LongProgress> Progress => progress.AsObservable();
+
+
+    public override string ToString()
+    {
+        return $"{nameof(Left)}: {Left}, {nameof(Source)}: {Source}, {nameof(Destination)}: {Destination}";
+    }
+
+    private Task<Result> OnSync(IZafiroDirectory source, IZafiroDirectory destination, CancellationToken ct)
+    {
+        return source.GetFromPath(Left).CombineAndBind(destination.GetFromPath(Left), (src, dst) =>
+        {
+            return CopyFileAction
+                .Create(src, dst)
+                .Bind(async action =>
+                {
+                    using (action.Progress.Subscribe(progress))
+                    {
+                        return await action.Execute(ct);
+                    }
+                });
+        });
+    }
+}
